LayoutTests/ietestcenter/Javascript/15.3.4.5-2-2-expected.txt

15.3.4.5-2-2 (TestCases/chapter15/15.3/15.3.4/15.3.4.5/15.3.4.5-2-2.js) Function.prototype.bind throws TypeError if the Target is not callable (bind attached to object)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL ES5Harness.preconditionPassed should be true (of type boolean). Was undefined (of type undefined).
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
